requirements.txt
appdirs==1.4.3
beautifulsoup4==4.6.0
certifi==2017.4.17
chardet==3.0.3
dj-database-url==0.4.2
Django==1.11
django-cors-headers==2.0.2
django-debug-toolbar==1.7
djangorestframework==3.6.2
idna==2.5
packaging==16.8
pyparsing==2.2.0
pytz==2017.2
requests==2.17.3
six==1.10.0
sqlparse==0.2.3
urllib3==1.21.1
